OV-125: * move player controls components to folders
