+++ frontend/src/bundles/studio/components/player-controls/components/control/control.tsx
