+++ frontend/src/bundles/studio/components/player-controls/components/time-display/time-display.tsx
